HierarchyCache optimization
Optimized the cache. After figuring out that recursion was timeconsuming,
changed the _sync function to process changes using a queue.
Further, after finding out that querying data from the stage is time
consuming, tried to avoid it as much as possible (for ex. by storing the
UsdPrim in the node). Also changed the cache change data structure from
sets to vectors.