docs: revision of the sphinx introduction
Reworked the intial introduction to provide a more verbose explanation of the
problem set that UNF is designed to address. Added more detail about how UNF
does not affect the internal working of Hydra or composition and makes no
efforts to improve the speed or internal workings of USD.  Added simple code
examples and shell output to demonstrate how UNF can be used to improve high
level APIs and tools that directly watch the USD stage for changes.

Signed-off-by: Matthew E. Levine <melevine@gmail.com>